cloud-functions/nlp-enricher/requirements.txt
functions-framework==3.4.0
google-cloud-aiplatform==1.36.4
google-cloud-bigquery==3.11.4
flask==2.3.3
